+++ cmd/create.go
@@ -170,9 +170,13 @@ func init() {
 	cmdCreatePG.AddCommand(cmdCreatePGBinding)
 
 	// create cluster a8s
-	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureRegion, "backup-region", "eu-central-1", "specify the infrastructure region to store backups such as \"us-east-1\".")
+	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureRegion, "backup-region", "us-east-1", "specify the infrastructure region to store backups such as \"us-east-1\".")
 	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureBucket, "backup-bucket", "a8s-backups", "specify the infrastructure object store bucket name.")
-	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureProvider, "backup-provider", "AWS", "specify the infrastructure provider as supported by the a8s Backup Manager.")
+	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureProvider, "backup-provider", "minio", "specify the infrastructure provider as supported by the a8s Backup Manager. Valid options are: minio and AWS.")
+	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureEndpoint, "backup-store-endpoint", "http://minio.minio-dev.svc.cluster.local:9000", "the endpoint of the S3 compatible backup object storage. Default is for non-prod Minio setups.")
+	cmdCreateClusterA8s.PersistentFlags().BoolVar(&demo.BackupInfrastructurePathStyle, "backup-store-pathstyle", true, "influences the URI schema used to talk to the S3 compatible backup object store. Set to false for S3 and true for Minio.")
+	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupStoreAccessKey, "backup-store-accesskey", "a8s-user", "the access key id for the backup store.")
+	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.BackupStoreSecretKey, "backup-store-secretkey", "a8s-password", "the secret key for the backup store.")
 	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.DeploymentVersion, "deployment-version", "v0.3.0", "specify the version corresponding to the a8s-deployment git version tag. Use \"latest\" to get the untagged version.")
 	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.ClusterNrOfNodes, "cluster-nr-of-nodes", "3", "specify number of Kubernetes nodes.")
 	cmdCreateClusterA8s.PersistentFlags().StringVar(&demo.ClusterMemory, "cluster-memory", "4gb", "specify memory of the Kubernetes cluster.")
@@ -181,7 +185,11 @@ func init() {
 	// create stack a8s
 	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureRegion, "backup-region", "eu-central-1", "specify the infrastructure region to store backups such as \"us-east-1\".")
 	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureBucket, "backup-bucket", "a8s-backups", "specify the infrastructure object store bucket name.")
-	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureProvider, "backup-provider", "AWS", "specify the infrastructure provider as supported by the a8s Backup Manager.")
+	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureProvider, "backup-provider", "minio", "specify the infrastructure provider as supported by the a8s Backup Manager. Valid options are: minio and AWS.")
+	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupInfrastructureEndpoint, "backup-store-endpoint", "http://minio.minio-dev.svc.cluster.local:9000", "the endpoint of the S3 compatible backup object storage. Default is for non-prod Minio setups.")
+	cmdCreateStackA8s.PersistentFlags().BoolVar(&demo.BackupInfrastructurePathStyle, "backup-store-pathstyle", true, "influences the URI schema used to talk to the S3 compatible backup object store. Set to false for S3 and true for Minio.")
+	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupStoreAccessKey, "backup-store-accesskey", "a8s-user", "the access key id for the backup store.")
+	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.BackupStoreSecretKey, "backup-store-secretkey", "a8s-password", "the secret key for the backup store.")
 	cmdCreateStackA8s.PersistentFlags().StringVar(&demo.DeploymentVersion, "deployment-version", "v0.3.0", "specify the version corresponding to the a8s-deployment git version tag. Use \"latest\" to get the untagged version.")
 	cmdCreateStackA8s.PersistentFlags().BoolVar(&demo.NoPreCheck, "no-precheck", false, "skip the verification of prerequisites.")