+++ demo/required_commands.go
@@ -1,5 +1,11 @@
 package demo
 
+/*
+List of CLI and other binaries required for using the a9s CLI.
+*/
+
+//var requiredCommands map[string]map[string]string
+
 func RequiredCommands() map[string]map[string]string {
 	cmds := make(map[string]map[string]string)
 
@@ -11,6 +17,9 @@ func RequiredCommands() map[string]map[string]string {
 
 	cmds["cmctl"] = make(map[string]string)
 	cmds["cmctl"]["darwin"] = "brew install cmctl"
+
+	cmds["mc"] = make(map[string]string)
+	cmds["cmctl"]["darwin"] = "brew install minio-mc"
 
 	return cmds
 }